Fix Driver Stats dto

+++ app/src/main/java/pe/edu/upc/flota365/features/manager/data/remote/models/DriverDto.kt
@@ -53,5 +53,5 @@ data class DriverStatsDto(
   val driversWithExpiredLicense: Int,
   val assignedDrivers: Int,
   val unassignedDrivers: Int,
-  val averageExperience: Int
+  val averageExperience: Double
 )
